refactor(gt7): reorganize code into focused modules
