+++ gt7/telemetry-server/Cargo.toml
@@ -22,3 +22,5 @@ openssl = { version = "0.10.59", features = ["vendored"] }
 [dev-dependencies]
 http-body-util = "0.1.3"
 tower = "0.5.2"
+tokio-tungstenite = "0.26"
+futures-util = "0.3"
