--- gt7/telemetry-server/tests/integration_tests.rs
@@ -1,19 +1,0 @@
-use std::env;
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_environment_variables_exist() {
-        // Test that we can read environment variables without panicking
-        let _ = env::var("PATH"); // This should always exist
-        assert!(true); // Basic test to ensure test framework works
-    }
-
-    #[test]
-    fn test_basic_functionality() {
-        // Test that basic functions don't panic
-        assert_eq!(2 + 2, 4);
-    }
-}
